api/requirements.txt
Flask==0.12.2
Flask-Login==0.4.0
Flask-Session==0.3.1
Flask_SocketIO
green
newsapi
patch>=1.16
pycryptodome>=3.4.3
pyjwt>=1.4.2
pylint
python-engineio
python-frontmatter
python-socketio
python-twitter
python-vote-full>=1.0
pyyaml>=3.11
redlock>=1.2.0
six==1.9.0
sqlalchemy
stem>=1.5.3
redis
unidiff
urwid>=1.3.1
Werkzeug==0.12.2
autopep8
